# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Entrepreneur Mitra Backend",
    "description": "Production-quality backend for SIH26092: AI-Driven Scheme Matching for Marginalized Entrepreneurs (Ministry of Social Justice and Empowerment - MoSJE). Connects conversational voice interview, deterministic rule evaluation, transparent ranking, financial affordability calculation, and geo-spatial partner routing.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/profiles": {
      "post": {
        "tags": [
          "Profiles"
        ],
        "summary": "Create Profile",
        "description": "Create a new entrepreneur profile (supports anonymous MVP or authenticated user).",
        "operationId": "create_profile_api_v1_profiles_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_ProfileOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/profiles/{profile_id}": {
      "get": {
        "tags": [
          "Profiles"
        ],
        "summary": "Get Profile By Id",
        "description": "Retrieve an entrepreneur profile by ID.",
        "operationId": "get_profile_by_id_api_v1_profiles__profile_id__get",
        "parameters": [
          {
            "name": "profile_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Profile Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_ProfileOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Profiles"
        ],
        "summary": "Update Profile By Id",
        "description": "Update profile attributes or preferred language.",
        "operationId": "update_profile_by_id_api_v1_profiles__profile_id__patch",
        "parameters": [
          {
            "name": "profile_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Profile Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_ProfileOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/profile": {
      "get": {
        "tags": [
          "Profiles"
        ],
        "summary": "Get My Profile",
        "description": "Get current profile by query param or authenticated user.",
        "operationId": "get_my_profile_api_v1_profile_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "profile_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Profile Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_ProfileOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/schemes": {
      "get": {
        "tags": [
          "Schemes"
        ],
        "summary": "List Schemes",
        "description": "List all verified government schemes with optional filtering.",
        "operationId": "list_schemes_api_v1_schemes_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": "ACTIVE",
              "title": "Status"
            }
          },
          {
            "name": "scheme_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Scheme Type"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_List_SchemeOut__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/schemes/{scheme_id}": {
      "get": {
        "tags": [
          "Schemes"
        ],
        "summary": "Get Scheme Details",
        "description": "Retrieve full details of a scheme including rules, benefits, and verified official sources.",
        "operationId": "get_scheme_details_api_v1_schemes__scheme_id__get",
        "parameters": [
          {
            "name": "scheme_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Scheme Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_SchemeDetailOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/schemes/{scheme_id}/application-steps": {
      "get": {
        "tags": [
          "Schemes"
        ],
        "summary": "Get Scheme Application Steps",
        "description": "Get authoritative application guidance and steps for a specific scheme.",
        "operationId": "get_scheme_application_steps_api_v1_schemes__scheme_id__application_steps_get",
        "parameters": [
          {
            "name": "scheme_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Scheme Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/saved-schemes": {
      "get": {
        "tags": [
          "Saved Schemes"
        ],
        "summary": "List Saved Schemes",
        "description": "List all bookmarked schemes.",
        "operationId": "list_saved_schemes_api_v1_saved_schemes_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_List_SavedSchemeOut__"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Saved Schemes"
        ],
        "summary": "Save Scheme",
        "description": "Bookmark/save a scheme for later access.",
        "operationId": "save_scheme_api_v1_saved_schemes_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveSchemeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/schemes/{scheme_id}/save": {
      "delete": {
        "tags": [
          "Saved Schemes"
        ],
        "summary": "Unsave Scheme",
        "description": "Remove a saved scheme bookmark.",
        "operationId": "unsave_scheme_api_v1_schemes__scheme_id__save_delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "scheme_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Scheme Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiResponse_List_SavedSchemeOut__": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "data": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/SavedSchemeOut"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorDetail"
              },
              {
                "type": "null"
              }
            ]
          },
          "meta": {
            "$ref": "#/components/schemas/MetaInfo"
          }
        },
        "type": "object",
        "title": "ApiResponse[List[SavedSchemeOut]]"
      },
...
      "ApiResponse_List_SchemeOut__": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "data": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/SchemeOut"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorDetail"
              },
              {
                "type": "null"
              }
            ]
          },
          "meta": {
            "$ref": "#/components/schemas/MetaInfo"
          }
        },
        "type": "object",
        "title": "ApiResponse[List[SchemeOut]]"
      },
...
      "ApiResponse_ProfileOut_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProfileOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorDetail"
              },
              {
                "type": "null"
              }
            ]
          },
          "meta": {
            "$ref": "#/components/schemas/MetaInfo"
          }
        },
        "type": "object",
        "title": "ApiResponse[ProfileOut]"
      },
      "ApiResponse_SchemeDetailOut_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SchemeDetailOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorDetail"
              },
              {
                "type": "null"
              }
            ]
          },
          "meta": {
            "$ref": "#/components/schemas/MetaInfo"
          }
        },
        "type": "object",
        "title": "ApiResponse[SchemeDetailOut]"
      },
...
      "ApiResponse_dict_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "data": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorDetail"
              },
              {
                "type": "null"
              }
            ]
          },
          "meta": {
            "$ref": "#/components/schemas/MetaInfo"
          }
        },
        "type": "object",
        "title": "ApiResponse[dict]"
      },
...
      "ErrorDetail": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "user_message": {
            "type": "string",
            "title": "User Message"
          },
          "retryable": {
            "type": "boolean",
            "title": "Retryable",
            "default": false
          },
          "details": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Details"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message",
          "user_message"
        ],
        "title": "ErrorDetail"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MetaInfo": {
        "properties": {
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "timestamp": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timestamp"
          }
        },
        "type": "object",
        "title": "MetaInfo"
      },
...
      "ProfileCreate": {
        "properties": {
          "language": {
            "type": "string",
            "title": "Language",
            "default": "hi"
          },
          "attributes": {
            "additionalProperties": true,
            "type": "object",
            "title": "Attributes"
          }
        },
        "type": "object",
        "title": "ProfileCreate"
      },
      "ProfileOut": {
        "properties": {
          "profile_id": {
            "type": "string",
            "title": "Profile Id"
          },
          "user_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "language": {
            "type": "string",
            "title": "Language"
          },
          "attributes": {
            "additionalProperties": true,
            "type": "object",
            "title": "Attributes"
          },
          "attribute_details": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/AttributeSchema"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Attribute Details"
          },
          "completeness_pct": {
            "type": "integer",
            "title": "Completeness Pct"
          },
          "confirmed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confirmed At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "profile_id",
          "status",
          "language",
          "attributes",
          "completeness_pct",
          "created_at",
          "updated_at"
        ],
        "title": "ProfileOut"
      },
      "ProfileUpdate": {
        "properties": {
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "attributes": {
            "additionalProperties": true,
            "type": "object",
            "title": "Attributes"
          }
        },
        "type": "object",
        "title": "ProfileUpdate"
      },
      "SaveSchemeRequest": {
        "properties": {
          "scheme_id": {
            "type": "string",
            "title": "Scheme Id"
          }
        },
        "type": "object",
        "required": [
          "scheme_id"
        ],
        "title": "SaveSchemeRequest"
      },
...
      "SchemeOut": {
        "properties": {
          "scheme_id": {
            "type": "string",
            "title": "Scheme Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "short_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Short Name"
          },
          "ministry": {
            "type": "string",
            "title": "Ministry"
          },
          "department": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Department"
          },
          "scheme_type": {
            "type": "string",
            "title": "Scheme Type"
          },
          "target_group": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Group"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "official_url": {
            "type": "string",
            "title": "Official Url"
          },
          "application_url": {
            "type": "string",
            "title": "Application Url"
          },
          "helpline": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Helpline"
          },
          "source_confidence": {
            "type": "string",
            "title": "Source Confidence"
          },
          "last_verified_at": {
            "type": "string",
            "format": "date-time",
            "title": "Last Verified At"
          }
        },
        "type": "object",
        "required": [
          "scheme_id",
          "name",
          "ministry",
          "scheme_type",
          "description",
          "status",
          "official_url",
          "application_url",
          "source_confidence",
          "last_verified_at"
        ],
        "title": "SchemeOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}